requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
SQLAlchemy>=2.0
fastapi>=0.100.0
uvicorn>=0.22.0
icalendar>=5.0.0
aiogram>=3.3.0
python-dotenv>=1.0.0
apscheduler>=3.10.1
